+++ src/components/common/Slider/SliderArrow.tsx
@@ -1,9 +1,0 @@
-import styles from './SliderArrow.module.scss';
-
-interface SliderArrowProps {
-  position: 'prev' | 'next';
-}
-
-export default function SliderArrow({ position }: SliderArrowProps) {
-  return <span className={`${styles.SliderArrow} ${styles[`SliderArrow_${position}`]}`}></span>;
-}