+++ src/components/common/Slider/index.ts
@@ -1,3 +1,0 @@
-import SliderArrow from './SliderArrow';
-
-export { SliderArrow };